refactor: renames "definitionWithAugmentation" to "definitionAugmentation"
