+++ src/library/Shortfin/TextToImage/SDXL/Client/Request/Body/definitionAugmentation.ts
