+++ src/library/Shortfin/TextToImage/SDXL/Client/Request/Body/exports.ts
@@ -1,4 +1,4 @@
-import './definitionWithAugmentation.ts';
+import './definitionAugmentation.ts';
 
 export type {
   Shortfin_TextToImage_SDXL_Client_Request_Body,
